IGNITE-21511 Make tests more readable
